Initial commit.
So far, this tool does the following:

1. Copies the pokeemerald directory to the output directory to use as a base.

2. Increases the number of total palettes used by maps from 13 to 14, allowing 1 more secondary palette per map.

For each map in Emerald, the first 6 palettes come from the Primary tileset, while the remainder come from the Secondary tileset. For FireRed the first 7 palettes are from the Primary tileset. When importing FireRed maps, any tiles referencing that 7th palette will instead use the 7th palette of the Secondary tileset. The plan is to take the "lost" 7th Primary palette and copy it to the new 14th palette of any Secondary tilesets that are paired with the current Primary tileset, and then update all the metatiles to change any references to the 7th palette to a reference to the 14th palette. This only works if every Secondary tileset is used exclusively with only one of the Primary tilesets. Any cases where a Secondary tileset is used with both Primary tilesets across all maps would require a duplicate Secondary tileset and palette, one for each Primary tileset.

3. Edits  \data\tilesets\header.h of the copied pokeemerald to add all the entries found in the pokefirered equivilent.

This process also does the following:
- Swaps the last two lines of each entry so the gMetatileAttributes is before the animation callback.
- Changes any "0x0" values to "NULL"
- Adds a "FR_" prefix to all names. (ex. "gTileset_FR_General") to avoid conflicts and to easily differentiate between Emerald and FireRed names. I may change this to a postfix instead, or only add it to names when there is a conflict.